+++ src/elements/DEM_ellip3d/ellip3d/root6.h
@@ -1,9 +1,10 @@
 #ifndef ROOT6_H
 #define ROOT6_H
 
+#include "realtypes.h"
 #include "vec.h"
 
 namespace dem {
-    bool root6(long double coef1[],long double coef2[],vec& v);
+    bool root6(REAL coef1[],REAL coef2[],vec& v);
 }
 #endif
